+++ packages/datadog_webview_tracking/example/android/app/src/main/kotlin/com/datadoghq/example/webviewTracking/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.datadoghq.example.webviewTracking
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()